// src/app/shared/model/app-referantor/app-referantor-tax-info-obj.model.ts
export class AppReferantorTaxInfoObj {
    TaxpayerNo: string;
    TaxIdNo: string;
    TaxIdName: string;
    TaxIdAddr: string;
    TaxIdAreaCode1: string;
    TaxIdAreaCode2: string;
    TaxIdAreaCode3: string;
    TaxIdAreaCode4: string;
    TaxIdCity: string;
    TaxIdZipcode: string;
  
    constructor() {
      this.TaxpayerNo = "";
      this.TaxIdNo = "";
      this.TaxIdName = "";
      this.TaxIdAddr = "";
      this.TaxIdAreaCode1 = "";
      this.TaxIdAreaCode2 = "";
      this.TaxIdAreaCode3 = "";
      this.TaxIdAreaCode4 = "";
      this.TaxIdCity = "";
      this.TaxIdZipcode = "";
    }
}